TST: highlight_tool_test_case.py
